backend/requirements.txt
alembic==1.13.2
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.65.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pydantic==1.8.2
pytest==8.2.2
pytest-asyncio==0.23.7
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.14.2
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
